changed sp-nav-offset to be percentage based instead of px based.
I change the @sp-nav-offset variable to 80% instead of 272px. Using percentage in a case like this allows you to fully adapt to the variable sizes of phone screens. Additionally, i think 80% is big enough to display additional information in the nav (if you were to use this as a menu, and not a nav) but still small enough to display the nav button